Moved the uniqueness classes to a separate file
