+++ src/core/layer.scala
@@ -56,7 +56,7 @@ case class Layer(version: Int,
 
   def localUniverse(hierarchy: Hierarchy, path: Pointer): Universe = Universe(
     hierarchy = hierarchy,
-    projects = projects.map { project => project.id -> Universe.Unique(project.projectRef, Set(path)) }.toMap,
+    projects = projects.map { project => project.id -> Uniqueness.Unique(project.projectRef, Set(path)) }.toMap,
     repoSets = repos.groupBy(_.commit.repoSetId).mapValues(_.map(_.ref(path))),
     imports = imports.map { i => i.layerRef.short -> LayerEntity(i.layerRef.short, Map(path -> i)) }.toMap
   )
